Split backend SQL parser modules
