+++ apps/backend/src/aiTools/sqlDialect/parser.ts
@@ -1,432 +1,36 @@
 import {
   ensureSqlSourceColumnExists,
-  getSqlColumnDescriptor,
   isSqlResourceName,
 } from "./schema";
+import {
+  parseDeleteStatement,
+  parseInsertStatement,
+  parseUpdateStatement,
+} from "./mutationParser";
+import {
+  parsePredicateClauses,
+  parseStringLiteral,
+} from "./predicateParser";
+import {
+  extractTopLevelClauses,
+  findTopLevelClauseMatches,
+  normalizeSqlWhitespace,
+  splitSqlStatements,
+  splitTopLevel,
+  upperCaseKeyword,
+} from "./parserSplitting";
 import type {
   ParsedSqlStatement,
   SqlAggregateFunctionName,
-  SqlComparisonOperator,
-  SqlDeleteStatement,
   SqlFromSource,
-  SqlInsertStatement,
-  SqlLiteral,
   SqlOrderDirection,
-  SqlPredicate,
-  SqlPredicateClause,
-  SqlPredicateValue,
   SqlSelectItem,
   SqlSelectOrderBy,
   SqlSelectStatement,
   SqlShowTablesStatement,
-  SqlUpdateStatement,
 } from "./types";
 
-function upperCaseKeyword(value: string): string {
-  return value.trim().replace(/\s+/g, " ").toUpperCase();
-}
-
-function normalizeSqlWhitespace(value: string): string {
-  const trimmedValue = value.trim();
-  let normalizedValue = "";
-  let inString = false;
-  let pendingWhitespace = false;
-
-  for (let index = 0; index < trimmedValue.length; index += 1) {
-    const character = trimmedValue[index];
-    const nextCharacter = trimmedValue[index + 1];
-
-    if (character === "'") {
-      if (pendingWhitespace && normalizedValue !== "") {
-        normalizedValue += " ";
-        pendingWhitespace = false;
-      }
-
-      normalizedValue += character;
-      if (inString && nextCharacter === "'") {
-        normalizedValue += nextCharacter;
-        index += 1;
-        continue;
-      }
-
-      inString = !inString;
-      continue;
-    }
-
-    if (inString) {
-      normalizedValue += character;
-      continue;
-    }
-
-    if (/\s/u.test(character)) {
-      pendingWhitespace = true;
-      continue;
-    }
-
-    if (character === ";" && trimmedValue.slice(index + 1).trim() === "") {
-      break;
-    }
-
-    if (pendingWhitespace && normalizedValue !== "") {
-      normalizedValue += " ";
-      pendingWhitespace = false;
-    }
-
-    normalizedValue += character;
-  }
-
-  return normalizedValue;
-}
-
-export function splitSqlStatements(value: string): ReadonlyArray<string> {
-  const trimmedValue = value.trim();
-  if (trimmedValue === "") {
-    return [];
-  }
-
-  const statements: Array<string> = [];
-  let current = "";
-  let inString = false;
-  let depth = 0;
-
-  for (let index = 0; index < trimmedValue.length; index += 1) {
-    const character = trimmedValue[index];
-    const nextCharacter = trimmedValue[index + 1];
-
-    if (character === "'") {
-      current += character;
-      if (inString && nextCharacter === "'") {
-        current += nextCharacter;
-        index += 1;
-      } else {
-        inString = !inString;
-      }
-      continue;
-    }
-
-    if (inString) {
-      current += character;
-      continue;
-    }
-
-    if (character === "(") {
-      depth += 1;
-      current += character;
-      continue;
-    }
-
-    if (character === ")") {
-      depth -= 1;
-      current += character;
-      continue;
-    }
-
-    if (depth === 0 && character === ";") {
-      const statement = current.trim();
-      const remainingValue = trimmedValue.slice(index + 1).trim();
-      if (statement === "") {
-        throw new Error("SQL batch contains an empty statement");
-      }
-
-      statements.push(statement);
-      current = "";
-
-      if (remainingValue === "") {
-        break;
-      }
-
-      continue;
-    }
-
-    current += character;
-  }
-
-  const statement = current.trim();
-  if (statement !== "") {
-    statements.push(statement);
-  }
-
-  return statements;
-}
-
-function assert(condition: boolean, message: string): void {
-  if (condition === false) {
-    throw new Error(message);
-  }
-}
-
-function splitTopLevel(value: string, separator: string): ReadonlyArray<string> {
-  const parts: Array<string> = [];
-  let current = "";
-  let inString = false;
-  let inDoubleQuote = false;
-  let depth = 0;
-
-  for (let index = 0; index < value.length; index += 1) {
-    const character = value[index];
-    const nextCharacter = value[index + 1];
-    if (character === "'" && inDoubleQuote === false) {
-      current += character;
-      if (inString && nextCharacter === "'") {
-        current += nextCharacter;
-        index += 1;
-      } else {
-        inString = !inString;
-      }
-      continue;
-    }
-
-    if (inString) {
-      current += character;
-      continue;
-    }
-
-    // Double-quoted segments appear inside Postgres text-array literals ({"a,b"}).
-    // Track them so a comma inside a quoted element is not treated as a separator.
-    if (character === '"') {
-      current += character;
-      const isEscapedQuote = inDoubleQuote && value[index - 1] === "\\";
-      if (isEscapedQuote === false) {
-        inDoubleQuote = !inDoubleQuote;
-      }
-
-      continue;
-    }
-
-    if (inDoubleQuote) {
-      current += character;
-      continue;
-    }
-
-    if (character === "(" || character === "[" || character === "{") {
-      depth += 1;
-      current += character;
-      continue;
-    }
-
-    if (character === ")" || character === "]" || character === "}") {
-      depth -= 1;
-      current += character;
-      continue;
-    }
-
-    if (depth === 0 && value.slice(index, index + separator.length).toUpperCase() === separator) {
-      parts.push(current.trim());
-      current = "";
-      index += separator.length - 1;
-      continue;
-    }
-
-    current += character;
-  }
-
-  if (current.trim() !== "") {
-    parts.push(current.trim());
-  }
-
-  return parts;
-}
-
-function splitTopLevelByKeyword(value: string, keyword: "AND" | "OR"): ReadonlyArray<string> {
-  const parts: Array<string> = [];
-  let current = "";
-  let inString = false;
-  let inDoubleQuote = false;
-  let depth = 0;
-
-  for (let index = 0; index < value.length; index += 1) {
-    const character = value[index];
-    const nextCharacter = value[index + 1];
-    if (character === "'" && inDoubleQuote === false) {
-      current += character;
-      if (inString && nextCharacter === "'") {
-        current += nextCharacter;
-        index += 1;
-      } else {
-        inString = !inString;
-      }
-      continue;
-    }
-
-    if (inString) {
-      current += character;
-      continue;
-    }
-
-    // Double-quoted segments appear inside Postgres text-array literals ({"a,b"}).
-    // Track them so a keyword inside a quoted element is not treated as a separator.
-    if (character === '"') {
-      current += character;
-      const isEscapedQuote = inDoubleQuote && value[index - 1] === "\\";
-      if (isEscapedQuote === false) {
-        inDoubleQuote = !inDoubleQuote;
-      }
-
-      continue;
-    }
-
-    if (inDoubleQuote) {
-      current += character;
-      continue;
-    }
-
-    if (character === "(" || character === "[" || character === "{") {
-      depth += 1;
-      current += character;
-      continue;
-    }
-
-    if (character === ")" || character === "]" || character === "}") {
-      depth -= 1;
-      current += character;
-      continue;
-    }
-
-    if (
-      depth === 0
-      && value.slice(index, index + keyword.length).toUpperCase() === keyword
-      && (index === 0 || /\s/.test(value[index - 1] ?? ""))
-      && (index + keyword.length >= value.length || /\s/.test(value[index + keyword.length] ?? ""))
-    ) {
-      parts.push(current.trim());
-      current = "";
-      index += keyword.length - 1;
-      continue;
-    }
-
-    current += character;
-  }
-
-  if (current.trim() !== "") {
-    parts.push(current.trim());
-  }
-
-  return parts;
-}
-
-type TopLevelClauseDefinition<TName extends string> = Readonly<{
-  name: TName;
-  keyword: string;
-}>;
-
-type TopLevelClauseMatch<TName extends string> = Readonly<{
-  name: TName;
-  keyword: string;
-  index: number;
-}>;
-
-function isSqlBoundaryCharacter(value: string | undefined): boolean {
-  return value === undefined || /\s/u.test(value);
-}
-
-function findTopLevelClauseMatches<TName extends string>(
-  value: string,
-  definitions: ReadonlyArray<TopLevelClauseDefinition<TName>>,
-): ReadonlyArray<TopLevelClauseMatch<TName>> {
-  const matches: Array<TopLevelClauseMatch<TName>> = [];
-  const normalizedDefinitions = [...definitions].sort((left, right) => right.keyword.length - left.keyword.length);
-  let inString = false;
-  let depth = 0;
-
-  for (let index = 0; index < value.length; index += 1) {
-    const character = value[index];
-    const nextCharacter = value[index + 1];
-    if (character === "'") {
-      if (inString && nextCharacter === "'") {
-        index += 1;
-        continue;
-      }
-
-      inString = !inString;
-      continue;
-    }
-
-    if (inString) {
-      continue;
-    }
-
-    if (character === "(") {
-      depth += 1;
-      continue;
-    }
-
-    if (character === ")") {
-      depth -= 1;
-      continue;
-    }
-
-    if (depth !== 0) {
-      continue;
-    }
-
-    const matchedDefinition = normalizedDefinitions.find((definition) => {
-      if (value.slice(index, index + definition.keyword.length).toUpperCase() !== definition.keyword) {
-        return false;
-      }
-
-      return isSqlBoundaryCharacter(value[index - 1]) && isSqlBoundaryCharacter(value[index + definition.keyword.length]);
-    });
-    if (matchedDefinition === undefined) {
-      continue;
-    }
-
-    matches.push({
-      name: matchedDefinition.name,
-      keyword: matchedDefinition.keyword,
-      index,
-    });
-    index += matchedDefinition.keyword.length - 1;
-  }
-
-  return matches;
-}
-
-function extractTopLevelClauses<TName extends string>(
-  value: string,
-  definitions: ReadonlyArray<TopLevelClauseDefinition<TName>>,
-  context: string,
-): Readonly<{
-  leadingSegment: string;
-  clauseValues: ReadonlyMap<TName, string>;
-}> {
-  const matches = findTopLevelClauseMatches(value, definitions);
-  if (matches.length === 0) {
-    return {
-      leadingSegment: value.trim(),
-      clauseValues: new Map(),
-    };
-  }
-
-  const definitionOrder = new Map(definitions.map((definition, index) => [definition.name, index] as const));
-  const clauseValues = new Map<TName, string>();
-  let lastOrder = -1;
-
-  for (const [index, match] of matches.entries()) {
-    if (clauseValues.has(match.name)) {
-      throw new Error(`Duplicate ${context} clause: ${match.keyword}`);
-    }
-
-    const order = definitionOrder.get(match.name);
-    if (order === undefined) {
-      throw new Error(`Unknown ${context} clause: ${match.keyword}`);
-    }
-    if (order < lastOrder) {
-      throw new Error(`Invalid ${context} clause order near ${match.keyword}`);
-    }
-
-    const nextMatch = matches[index + 1];
-    const clauseValue = value.slice(match.index + match.keyword.length, nextMatch?.index).trim();
-    clauseValues.set(match.name, clauseValue);
-    lastOrder = order;
-  }
-
-  const firstMatch = matches[0];
-  assert(firstMatch !== undefined, `Expected at least one ${context} clause`);
-  return {
-    leadingSegment: value.slice(0, firstMatch.index).trim(),
-    clauseValues,
-  };
-}
+export { splitSqlStatements } from "./parserSplitting";
 
 function parseSimpleNumberClauseValue(value: string | undefined, keyword: string): number | null {
   if (value === undefined) {
@@ -439,265 +43,6 @@ function parseSimpleNumberClauseValue(value: string | undefined, keyword: string
   }
 
   return Number.parseInt(trimmedValue, 10);
-}
-
-function parseStringLiteral(value: string): string {
-  assert(value.startsWith("'") && value.endsWith("'"), "Expected a quoted string literal");
-  return value.slice(1, -1).replaceAll("''", "'");
-}
-
-function parseSqlLiteral(value: string): SqlLiteral {
-  const trimmedValue = value.trim();
-  if (trimmedValue.toUpperCase() === "NULL") {
-    return null;
-  }
-
-  if (trimmedValue.toUpperCase() === "TRUE") {
-    return true;
-  }
-
-  if (trimmedValue.toUpperCase() === "FALSE") {
-    return false;
-  }
-
-  if (trimmedValue.startsWith("'") && trimmedValue.endsWith("'")) {
-    return parseStringLiteral(trimmedValue);
-  }
-
-  if (/^-?\d+$/.test(trimmedValue)) {
-    return Number.parseInt(trimmedValue, 10);
-  }
-
-  if (/^-?\d+\.\d+$/.test(trimmedValue)) {
-    return Number.parseFloat(trimmedValue);
-  }
-
-  throw new Error(`Unsupported literal: ${trimmedValue}`);
-}
-
-function parsePredicateValue(value: string): SqlPredicateValue {
-  const trimmedValue = value.trim();
-  if (trimmedValue.toUpperCase() === "NOW()") {
-    return { type: "now" };
-  }
-
-  return parseSqlLiteral(trimmedValue);
-}
-
-function parsePostgresTextArrayElement(item: string, columnName: string, trimmedValue: string): string {
-  const trimmedItem = item.trim();
-  if (trimmedItem.startsWith('"') && trimmedItem.endsWith('"') && trimmedItem.length >= 2) {
-    return trimmedItem.slice(1, -1).replaceAll('\\"', '"');
-  }
-
-  if (trimmedItem === "" || trimmedItem.includes('"')) {
-    throw new Error(
-      `Array column "${columnName}" accepts only string elements, e.g. {a,b} or {"a","b"}. Got: ${trimmedValue}`,
-    );
-  }
-
-  return trimmedItem;
-}
-
-function parseStringArrayLiteralList(value: string, columnName: string): ReadonlyArray<string> {
-  const trimmedValue = value.trim();
-
-  const arrayPrefixMatch = trimmedValue.match(/^ARRAY\s*\[([\s\S]*)\]$/i);
-  const bracketMatch = trimmedValue.startsWith("[") && trimmedValue.endsWith("]");
-  const parenMatch = trimmedValue.startsWith("(") && trimmedValue.endsWith(")");
-  const quotedBraceMatch = trimmedValue.startsWith("'{") && trimmedValue.endsWith("}'");
-  const braceMatch = trimmedValue.startsWith("{") && trimmedValue.endsWith("}");
-
-  // Postgres text-array literal: '{a,b}' or {a,b}, elements may be unquoted.
-  if (quotedBraceMatch || braceMatch) {
-    const innerValue = quotedBraceMatch ? trimmedValue.slice(2, -2).trim() : trimmedValue.slice(1, -1).trim();
-    if (innerValue === "") {
-      return [];
-    }
-
-    return splitTopLevel(innerValue, ",").map((item) =>
-      parsePostgresTextArrayElement(item, columnName, trimmedValue),
-    );
-  }
-
-  // Native ('a','b'), JSON-style ['a','b'], and ARRAY['a','b'] all unwrap to a
-  // comma-separated list of quoted string literals.
-  let innerValue: string | null = null;
-  if (arrayPrefixMatch !== null) {
-    innerValue = (arrayPrefixMatch[1] ?? "").trim();
-  } else if (bracketMatch || parenMatch) {
-    innerValue = trimmedValue.slice(1, -1).trim();
-  }
-
-  if (innerValue === null) {
-    throw new Error(
-      `Array column "${columnName}" expects a parenthesized list like ('tag1','tag2'), or () for an empty list. Got: ${trimmedValue}`,
-    );
-  }
-
-  if (innerValue === "") {
-    return [];
-  }
-
-  return splitTopLevel(innerValue, ",").map((item) => {
-    const parsedValue = parseSqlLiteral(item);
-    if (typeof parsedValue !== "string") {
-      throw new Error(
-        `Array column "${columnName}" accepts only quoted string literals, e.g. ('a','b'). Got: ${trimmedValue}`,
-      );
-    }
-
-    return parsedValue;
-  });
-}
-
-function parseLoweredStringLiteralList(value: string, operator: "IN" | "NOT IN"): ReadonlyArray<string> {
-  const trimmedValue = value.trim();
-  assert(trimmedValue.startsWith("(") && trimmedValue.endsWith(")"), "Expected a parenthesized value list");
-  const innerValue = trimmedValue.slice(1, -1).trim();
-  if (innerValue === "") {
-    return [];
-  }
-
-  return splitTopLevel(innerValue, ",").map((item) => {
-    const parsedValue = parseSqlLiteral(item);
-    if (typeof parsedValue !== "string") {
-      throw new Error(`LOWER(column) ${operator} (...) only supports string literals`);
-    }
-
-    return parsedValue;
-  });
-}
-
-function parsePredicate(source: SqlFromSource, value: string): SqlPredicate {
-  const trimmedValue = value.trim();
-  const matchPredicate = trimmedValue.match(/^MATCH\s*\(\s*('(?:''|[^'])*')\s*\)$/i);
-  if (matchPredicate !== null) {
-    return {
-      type: "match",
-      query: parseStringLiteral(matchPredicate[1] ?? ""),
-    };
-  }
-
-  const loweredLikePredicate = trimmedValue.match(/^LOWER\s*\(\s*([a-z_][a-z0-9_]*)\s*\)\s+LIKE\s+('(?:''|[^'])*')$/i);
-  if (loweredLikePredicate !== null) {
-    const columnName = (loweredLikePredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "like",
-      columnName,
-      pattern: parseStringLiteral(loweredLikePredicate[2] ?? ""),
-      caseInsensitive: true,
-    };
-  }
-
-  const loweredEqualsPredicate = trimmedValue.match(/^LOWER\s*\(\s*([a-z_][a-z0-9_]*)\s*\)\s*=\s*('(?:''|[^'])*')$/i);
-  if (loweredEqualsPredicate !== null) {
-    const columnName = (loweredEqualsPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "like",
-      columnName,
-      pattern: parseStringLiteral(loweredEqualsPredicate[2] ?? ""),
-      caseInsensitive: true,
-    };
-  }
-
-  const loweredInPredicate = trimmedValue.match(
-    /^LOWER\s*\(\s*([a-z_][a-z0-9_]*)\s*\)\s+(NOT\s+IN|IN)\s*(\(.+\))$/i,
-  );
-  if (loweredInPredicate !== null) {
-    const columnName = (loweredInPredicate[1] ?? "").toLowerCase();
-    const operator = (loweredInPredicate[2] ?? "").toUpperCase().replace(/\s+/g, " ");
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "in",
-      columnName,
-      values: parseLoweredStringLiteralList(loweredInPredicate[3] ?? "", operator as "IN" | "NOT IN"),
-      caseInsensitive: true,
-      isNegated: operator === "NOT IN",
-    };
-  }
-
-  const likePredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s+LIKE\s+('(?:''|[^'])*')$/i);
-  if (likePredicate !== null) {
-    const columnName = (likePredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "like",
-      columnName,
-      pattern: parseStringLiteral(likePredicate[2] ?? ""),
-      caseInsensitive: false,
-    };
-  }
-
-  const isNotNullPredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s+IS\s+NOT\s+NULL$/i);
-  if (isNotNullPredicate !== null) {
-    const columnName = (isNotNullPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "is_not_null",
-      columnName,
-    };
-  }
-
-  const isNullPredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s+IS\s+NULL$/i);
-  if (isNullPredicate !== null) {
-    const columnName = (isNullPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "is_null",
-      columnName,
-    };
-  }
-
-  const overlapPredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s+OVERLAP\s*(.+)$/i);
-  if (overlapPredicate !== null) {
-    const columnName = (overlapPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "overlap",
-      columnName,
-      values: parseStringArrayLiteralList(overlapPredicate[2] ?? "", columnName),
-    };
-  }
-
-  const inPredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s+IN\s*(\(.+\))$/i);
-  if (inPredicate !== null) {
-    const columnName = (inPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "in",
-      columnName,
-      values: splitTopLevel((inPredicate[2] ?? "").slice(1, -1), ",").map(parseSqlLiteral),
-      caseInsensitive: false,
-      isNegated: false,
-    };
-  }
-
-  const comparisonPredicate = trimmedValue.match(/^([a-z_][a-z0-9_]*)\s*(=|<=|>=|<|>)\s*(.+)$/i);
-  if (comparisonPredicate !== null) {
-    const columnName = (comparisonPredicate[1] ?? "").toLowerCase();
-    ensureSqlSourceColumnExists(source, columnName);
-    return {
-      type: "comparison",
-      columnName,
-      operator: (comparisonPredicate[2] ?? "=") as SqlComparisonOperator,
-      value: parsePredicateValue(comparisonPredicate[3] ?? ""),
-    };
-  }
-
-  throw new Error(`Unsupported predicate: ${trimmedValue}`);
-}
-
-function parsePredicateClauses(source: SqlFromSource, value: string): ReadonlyArray<SqlPredicateClause> {
-  const trimmedValue = value.trim();
-  if (trimmedValue === "") {
-    return [];
-  }
-
-  return splitTopLevelByKeyword(trimmedValue, "OR").map((clause) =>
-    splitTopLevelByKeyword(clause, "AND").map((predicate) => parsePredicate(source, predicate)));
 }
 
 function parseOrderBy(value: string): ReadonlyArray<SqlSelectOrderBy> {
@@ -931,166 +276,6 @@ function parseDescribeStatement(normalizedSql: string): ParsedSqlStatement | nul
   return {
     type: "describe",
     resourceName,
-    normalizedSql,
-  };
-}
-
-function parseInsertStatement(normalizedSql: string): SqlInsertStatement {
-  const match = normalizedSql.match(/^INSERT\s+INTO\s+([a-z_][a-z0-9_]*)\s*\((.+)\)\s+VALUES\s+([\s\S]+)$/i);
-  if (match === null) {
-    throw new Error(
-      "INSERT must list columns explicitly, e.g. INSERT INTO cards (front_text, back_text, tags) VALUES ('Q?', 'A', ('tag')). Array columns use a parenthesized list; () means empty.",
-    );
-  }
-
-  const resourceName = (match[1] ?? "").toLowerCase();
-  if (resourceName !== "cards" && resourceName !== "decks") {
-    throw new Error(`INSERT is not supported for ${resourceName}`);
-  }
-
-  const columnNames = splitTopLevel(match[2] ?? "", ",").map((columnName) => {
-    const normalizedColumnName = columnName.trim().toLowerCase();
-    const columnDescriptor = getSqlColumnDescriptor(resourceName, normalizedColumnName);
-    if (columnDescriptor.readOnly) {
-      throw new Error(`Column is read-only: ${normalizedColumnName}`);
-    }
-
-    return normalizedColumnName;
-  });
-
-  const rows = splitTopLevel(match[3] ?? "", ",").map((row) => row.trim()).filter((row) => row.startsWith("("));
-  assert(rows.length > 0, "INSERT must include at least one VALUES row");
-
-  const parsedRows = rows.map((row) => {
-    assert(
-      row.startsWith("(") && row.endsWith(")"),
-      `Invalid VALUES row: each row must be wrapped in parentheses, e.g. ('Q?', 'A', ('tag')). Got: ${row}`,
-    );
-    const values = splitTopLevel(row.slice(1, -1), ",").map((value, index) => {
-      const columnName = columnNames[index];
-      if (columnName === undefined) {
-        throw new Error(
-          `VALUES row contains more values than the ${columnNames.length} declared column(s) (${columnNames.join(", ")}). Got: ${row}`,
-        );
-      }
-
-      const columnDescriptor = getSqlColumnDescriptor(resourceName, columnName);
-      if (columnDescriptor.type === "string[]") {
-        return parseStringArrayLiteralList(value, columnName);
-      }
-
-      return parseSqlLiteral(value);
-    });
-
-    if (values.length !== columnNames.length) {
-      throw new Error(
-        `VALUES row does not match the ${columnNames.length} declared column(s) (${columnNames.join(", ")}); got ${values.length} value(s) in: ${row}`,
-      );
-    }
-
-    return values;
-  });
-
-  return {
-    type: "insert",
-    resourceName,
-    columnNames,
-    rows: parsedRows,
-    normalizedSql,
-  };
-}
-
-function parseAssignments(resourceName: "cards" | "decks", value: string): SqlUpdateStatement["assignments"] {
-  return splitTopLevel(value, ",").map((assignment) => {
-    const match = assignment.match(/^([a-z_][a-z0-9_]*)\s*=\s*([\s\S]+)$/i);
-    if (match === null) {
-      throw new Error(`Unsupported assignment: ${assignment}`);
-    }
-
-    const columnName = (match[1] ?? "").toLowerCase();
-    const columnDescriptor = getSqlColumnDescriptor(resourceName, columnName);
-    if (columnDescriptor.readOnly) {
-      throw new Error(`Column is read-only: ${columnName}`);
-    }
-
-    return {
-      columnName,
-      value: columnDescriptor.type === "string[]"
-        ? parseStringArrayLiteralList(match[2] ?? "", columnName)
-        : parseSqlLiteral(match[2] ?? ""),
-    };
-  });
-}
-
-function parseUpdateStatement(normalizedSql: string): SqlUpdateStatement {
-  const match = normalizedSql.match(/^UPDATE\s+([a-z_][a-z0-9_]*)([\s\S]*)$/i);
-  if (match === null) {
-    throw new Error("Unsupported UPDATE statement");
-  }
-
-  const resourceName = (match[1] ?? "").toLowerCase();
-  if (resourceName !== "cards" && resourceName !== "decks") {
-    throw new Error(`UPDATE is not supported for ${resourceName}`);
-  }
-
-  const source: SqlFromSource = {
-    resourceName,
-    unnestColumnName: null,
-    unnestAlias: null,
-  };
-  const extractedClauses = extractTopLevelClauses(
-    (match[2] ?? "").trim(),
-    [
-      { name: "set", keyword: "SET" },
-      { name: "where", keyword: "WHERE" },
-    ] as const,
-    "UPDATE",
-  );
-  const assignmentsValue = extractedClauses.clauseValues.get("set");
-  const predicateValue = extractedClauses.clauseValues.get("where");
-  if (extractedClauses.leadingSegment !== "" || assignmentsValue === undefined || predicateValue === undefined) {
-    throw new Error("Unsupported UPDATE statement");
-  }
-
-  return {
-    type: "update",
-    resourceName,
-    assignments: parseAssignments(resourceName, assignmentsValue),
-    predicateClauses: parsePredicateClauses(source, predicateValue),
-    normalizedSql,
-  };
-}
-
-function parseDeleteStatement(normalizedSql: string): SqlDeleteStatement {
-  const match = normalizedSql.match(/^DELETE\s+FROM\s+([a-z_][a-z0-9_]*)([\s\S]*)$/i);
-  if (match === null) {
-    throw new Error("Unsupported DELETE statement");
-  }
-
-  const resourceName = (match[1] ?? "").toLowerCase();
-  if (resourceName !== "cards" && resourceName !== "decks") {
-    throw new Error(`DELETE is not supported for ${resourceName}`);
-  }
-
-  const source: SqlFromSource = {
-    resourceName,
-    unnestColumnName: null,
-    unnestAlias: null,
-  };
-  const extractedClauses = extractTopLevelClauses(
-    (match[2] ?? "").trim(),
-    [{ name: "where", keyword: "WHERE" }] as const,
-    "DELETE",
-  );
-  const predicateValue = extractedClauses.clauseValues.get("where");
-  if (extractedClauses.leadingSegment !== "" || predicateValue === undefined) {
-    throw new Error("Unsupported DELETE statement");
-  }
-
-  return {
-    type: "delete",
-    resourceName,
-    predicateClauses: parsePredicateClauses(source, predicateValue),
     normalizedSql,
   };
 }
